consulta lista de jogadores finalizado

+++ player/src/main/java/com/sevenine/player/application/usecase/output/PagenateOutput.java
@@ -1,4 +1,4 @@
 package com.sevenine.player.application.usecase.output;
 
-public record PagenateOutput<T>(Object content, int page, int size, int totalPages) {
+public record PagenateOutput<T>(Object content, int page, int size, long totalPages) {
 }
